gdal2tiles.py - moved into osgeo_utils and added a wrapper script
